Improve style
