+++ src/main/scala/mailo/package.scala
@@ -1,55 +1,22 @@
 package mailo
 
-import com.typesafe.scalalogging.LazyLogging
-
-import mailo.http.MailClient
-import mailo.data.MailData
-import mailo.parser.HTMLParser
-
-import scala.concurrent.Future
-import scala.concurrent.ExecutionContext
+import java.io.{PrintWriter, StringWriter}
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
-
 import cats.data.EitherT
 import cats.instances.future._
-
 import com.typesafe.config.{Config, ConfigFactory}
-
-import akka.http.scaladsl.model.ContentType
+import com.typesafe.scalalogging.LazyLogging
+import mailo.MailRefinedContent._
+import mailo.data.{MailData, S3MailData}
+import mailo.http.{MailClient, MailgunClient}
+import mailo.parser.HTMLParser
+import scalacache._
+import scalacache.guava._
 
 import scala.concurrent.duration._
-import scalacache._
-import guava._
-
-import java.io.{PrintWriter, StringWriter}
-
-case class Attachment(
-  name: String,
-  `type`: ContentType,
-  content: String,
-  transferEncoding: Option[String] = None
-)
-
-case class MailResponse(
-  id: String,
-  message: String
-)
-
-case class MailRawContent(
-  template: String,
-  partials: Map[String, String]
-)
-
-object MailRefinedContent {
-  sealed abstract class MailRefinedContent(val content: String)
-
-  case class HTMLContent(override val content: String) extends MailRefinedContent(content)
-  case class TEXTContent(override val content: String) extends MailRefinedContent(content)
-}
-
-abstract class MailError(message: String) extends RuntimeException(message)
+import scala.concurrent.{ExecutionContext, Future}
 
 class Mailo(
   mailData: MailData,
@@ -59,8 +26,6 @@ class Mailo(
   ec: ExecutionContext,
   conf: Config = ConfigFactory.load()
 ) extends LazyLogging {
-  import MailRefinedContent._
-
   implicit private[this] val scalaCache = ScalaCache(GuavaCache())
 
   private[this] case class MailoConfig(cachingTTLSeconds: Int)
@@ -113,9 +78,6 @@ class S3MailgunMailo(
   materializer: ActorMaterializer,
   ec: ExecutionContext
 ) {
-  import data.S3MailData
-  import http.MailgunClient
-
   private[this] val s3 = new S3MailData()
   private[this] val mailgun = new MailgunClient()
