+++ src/main/scala/mailo/parser/TemplateParser.scala
@@ -1,46 +1,15 @@
 package mailo.parser
 
-import cats.syntax.either._
-import mailo.MailError
-import mailo.MailRawContent
-import util.matching.Regex
+import java.util.regex.Matcher
 
-object ParserError {
-  case object HtmlNotValid extends MailError("Content was not valid HTML")
-  case class DisjointParametersAndMatches(
-    justParams: Set[String],
-    justMatches: Set[String]
-  ) extends MailError(
-        "Disjoint parameters and matches, params (${justParams.toString}), matches (${jusetMatches.toString})"
-      )
-  case class OverlappedParametersAndMatches(
-    justParams: Set[String],
-    justMatches: Set[String],
-    overlap: Set[String]
-  ) extends MailError(
-        s"Overlapped parameters and matches, but no exact match: just params (${justParams.toString}), just matches (${justMatches.toString}), overlapped params (${overlap.toString})"
-      )
-  case class PartialsDoNotExist(partials: Set[String])
-      extends MailError(
-        s"Some of the provided partials do not exist, here is the list ${partials.toString}"
-      )
-  case class TooFewPartialsProvided(partials: Set[String])
-      extends MailError(
-        s"Too few partials provided to the document, here is the list ${partials.toString}"
-      )
-  case class TooFewParamsProvided(params: Set[String])
-      extends MailError(
-        s"Too few params provided to the document, here is the list ${params.toString}"
-      )
-  case class TooManyParamsProvided(params: Set[String])
-      extends MailError(
-        s"Too many params provided to the document, here is the list ${params.toString}"
-      )
-}
+import cats.syntax.either._
+import mailo.parser.ParserError._
+import mailo.{MailError, MailRawContent}
+
+import scala.util.matching.Regex
+import scala.util.matching.Regex.Match
 
 object HTMLParser {
-  import ParserError._
-
   def parse(content: MailRawContent, params: Map[String, String]): Either[MailError, String] =
     replaceAllPartials(content.template, content.partials).flatMap(replaceAllParams(_, params))
 
@@ -92,9 +61,7 @@ object HTMLParser {
     values: Map[String, String],
     pattern: Regex
   ): String = {
-    import scala.util.matching.Regex.Match
     def replacement(m: Match): String = {
-      import java.util.regex.Matcher
       require(m.groupCount == 1)
       Matcher.quoteReplacement(values(m.group(1)))
     }
